requirement.txt
openai>=1.30.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
numpy
requests
beautifulsoup4
lxml
sentence-transformers>=2.2.0
streamlit>=1.30.0